Initialize deck per the instructions depending on num players.
Adds a new helper module: `paxpamir_util.php` for game logic helpers.

New method `getInitialDeck` builds a new randomized deck according to
the instructions in the manual, forming the piles of various card types,
shuffling them, and then combining them together.  The method returns
plain card numbers, which are then used to build the token set for the
cards used in the game.

Cards not included in the initial deck will not be used for the rest of
the game, and so do not need to be added to the token collection.